wip: translation fixes
